cleanups

+++ gpu/nvda/include/dnnsp2.cuh
@@ -3,18 +3,16 @@ typedef enum {
     no
 } refine_t; 
 
-
 typedef enum {
     fp64,
     fp32,
     fp16_fp32
 } precision_t; 
 
-
 void dnnsp2(double*, 
             double*, 
-            size_t, 
-            size_t,
+            int, 
+            int,
             precision_t,
             refine_t);
